squashing non-isolated classes, adding to credits
